751 : Review fixes
